requirements.txt
langchain==0.2.15
langchain-community==0.2.15
langchain-openai==0.1.23
langchain-astradb>=0.3.0  # Let langchain-astradb pick compatible astrapy version
astrapy>=1.2,<2.0  # Compatible with langchain-astradb >= 0.3.0
python-dotenv
django
pypdf
daphne
channels
djangorestframework
django-cors-headers
django-extensions
transformers
torch
psycopg2
pgvector
jupyter
ipykernel
requests_cache>=0.9
geopy>=2.2.0
pandas>=1.3.0
openmeteo-requests
retry_requests>=1.0.0
psycopg2>=2.9.0  # If using PostgreSQL; remove if using another DB
retry